# SortedDictionary Copy optimization