--- a/src/libraries/System.Collections/tests/Generic/SortedDictionary/SortedDictionary.Generic.Tests.cs
+++ b/src/libraries/System.Collections/tests/Generic/SortedDictionary/SortedDictionary.Generic.Tests.cs
@@ -57,21 +57,18 @@ namespace System.Collections.Tests
         {
             IComparer<TKey> comparer = GetKeyIComparer();
             IDictionary<TKey, TValue> source = GenericIDictionaryFactory(count);
-            SortedDictionary<TKey, TValue> copied = new SortedDictionary<TKey, TValue>(source, comparer);
-            Assert.Equal(source, copied);
-            Assert.Equal(comparer, copied.Comparer);
-        }
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void SortedDictionary_Generic_Constructor_SortedDictionary_IComparer(int count)
-        {
-            IComparer<TKey> comparer = GetKeyIComparer();
-            IDictionary<TKey, TValue> source = GenericIDictionaryFactory(count);
             SortedDictionary<TKey, TValue> sourceSorted = new SortedDictionary<TKey, TValue>(source, comparer);
+            Assert.Equal(source, sourceSorted);
+            Assert.Equal(comparer, sourceSorted.Comparer);
+            // Test copying a sorted dictionary.
             SortedDictionary<TKey, TValue> copied = new SortedDictionary<TKey, TValue>(sourceSorted, comparer);
             Assert.Equal(sourceSorted, copied);
             Assert.Equal(comparer, copied.Comparer);
+            // Test copying a sorted dictionary with a different comparer.
+            IComparer<TKey> reverseComparer = Comparer<TKey>.Create((key1, key2) => -comparer.Compare(key1, key2));
+            SortedDictionary<TKey, TValue> copiedReverse = new SortedDictionary<TKey, TValue>(sourceSorted, reverseComparer);
+            Assert.Equal(sourceSorted, copiedReverse);
+            Assert.Equal(reverseComparer, copiedReverse.Comparer);
         }
 
         #endregion